requirements.txt
phidata
python-dotenv
arxiv
googlesearch-python
wikipedia
pycountry
pubmed 
packaging
duckduckgo-search
pgvector
psycopg[binary]
pypdf
sqlalchemy
streamlit
google-generativeai
fastapi
groq
deepseek
openai
uvicorn